+++ Remnants/Behaviours/RemnantItemsLocationsBehaviour.cs
@@ -1,20 +1,23 @@
-﻿using LethalLib.Modules;
+﻿using HarmonyLib;
+using LethalLib.Modules;
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
 namespace Remnants.Behaviours
 {
-    public class RegisterItemLocationsBehaviour
+    public class RemnantItemsLocationsBehaviour
     {
 
         #region Variables
         private bool _hasInitialized = false;
         private bool _isRegistering = false;
-        private ObjectLocationData _shipObjectLocation = new ObjectLocationData(new string[] { "Environment", "HangarShip" }, null, true);
-        private ObjectLocationData _propObjectLocation = new ObjectLocationData(new string[] { "Environment", "Props" }, null, false);
-        private ObjectLocationData _rootObjectLocation = new ObjectLocationData(new string[] { }, null, false);
+        private static ObjectLocationData _shipObjectLocation = new ObjectLocationData(new string[] { "Environment", "HangarShip" }, null, true);
+        private static ObjectLocationData  _propObjectLocation = new ObjectLocationData(new string[] { "Environment", "Props" }, null, false);
+        private static ObjectLocationData _rootObjectLocation = new ObjectLocationData(new string[] { }, null, false);
+        private static List<string> _bannedItemsFromSaving = new List<string>();
         public ObjectLocationData ShipObjectLocation
         {
             get { return _shipObjectLocation; }
@@ -50,6 +53,7 @@ namespace Remnants.Behaviours
             {
                 _hasInitialized = true;
                 SceneManager.sceneLoaded += RegisterItemsLocations;
+                _bannedItemsFromSaving = Remnants.Instance.RemnantsConfig.GetBannedFromSavingItemNames();
             }
         }
         #endregion
@@ -76,6 +80,7 @@ namespace Remnants.Behaviours
             return remnantItemsArray;
         }
 
+
         public GrabbableObject[] GetItemsInProps()
         {
             GrabbableObject[] remnantItemsArray = null;
@@ -90,6 +95,59 @@ namespace Remnants.Behaviours
             GameObject[] grabbableObjects = rootGameObjects.Where(gmObject => gmObject.GetComponent<GrabbableObject>() != null).ToArray(); ;
             remnantItemsArray = grabbableObjects.Select(gameObj => gameObj.GetComponent<GrabbableObject>()).ToArray();
             return remnantItemsArray;
+        }
+
+
+
+        public GrabbableObject[] GetAllItems()
+        {
+            List<GrabbableObject> itemsList = new List<GrabbableObject>();
+            //Get items from ship location
+            GrabbableObject[] shipItemArray = _shipObjectLocation.LocationObject.GetComponentsInChildren<GrabbableObject>();
+            //Get items from root location
+            GameObject[] rootObjectItemArray = SceneManager.GetActiveScene().GetRootGameObjects().Where(gmObject => gmObject.GetComponent<GrabbableObject>() != null).ToArray();
+            GrabbableObject[] rootItemArray = rootObjectItemArray.Select(gmObject => gmObject.GetComponent<GrabbableObject>()).ToArray();
+            //Get items in prop location
+            GrabbableObject[] propItemsArray = _propObjectLocation.LocationObject.GetComponentsInChildren<GrabbableObject>();
+            itemsList.AddRange(shipItemArray);
+            itemsList.AddRange(rootItemArray);
+            itemsList.AddRange(propItemsArray);
+            return itemsList.ToArray();
+
+        }
+
+
+        public GrabbableObject[] GetItemsToSave()
+        {
+            var mls = Remnants.Instance.Mls;
+            if(Remnants.Instance.RemnantsConfig.ShouldSaveRemnantItems.Value == false)
+            {
+                mls.LogInfo("Not saving remnant items.");
+                //Get non remnant items from normal place
+                GrabbableObject[] nonRemnantItemsArray = UnityEngine.Object.FindObjectsByType<GrabbableObject>(FindObjectsInactive.Exclude, FindObjectsSortMode.None);
+                return nonRemnantItemsArray;
+            }
+            mls.LogInfo("Saving all items to save.");
+            List<GrabbableObject> itemsList = new List<GrabbableObject>();
+            //Get items from ship location
+            GrabbableObject[] shipItemArray = _shipObjectLocation.LocationObject.GetComponentsInChildren<GrabbableObject>();
+            List<GrabbableObject> shipItemList = shipItemArray.Where(grabObj =>
+            !(grabObj is RagdollGrabbableObject) && ((grabObj.isInShipRoom && grabObj.isInElevator) || (grabObj.isInElevator && grabObj.isPocketed))).ToList();
+            //Get items from root location
+            GameObject[] rootObjectItemArray = SceneManager.GetActiveScene().GetRootGameObjects().Where(gmObject => gmObject.GetComponent<GrabbableObject>() != null).ToArray();
+            GrabbableObject[] rootItemArray = rootObjectItemArray.Select(gmObject => gmObject.GetComponent<GrabbableObject>()).ToArray();
+            List<GrabbableObject> rootItemList = rootItemArray.Where(grabObj =>
+            !(grabObj is RagdollGrabbableObject) && ((grabObj.isInShipRoom && grabObj.isInElevator) || (grabObj.isInElevator && grabObj.isPocketed))).ToList();
+            //Get items in prop location
+            GrabbableObject[] propItemsArray = _propObjectLocation.LocationObject.GetComponentsInChildren<GrabbableObject>();
+            List<GrabbableObject> propItemsList = propItemsArray.Where(grabObj =>
+            !(grabObj is RagdollGrabbableObject) && ((grabObj.isInShipRoom && grabObj.isInElevator) || (grabObj.isInElevator && grabObj.isPocketed))).ToList();
+            itemsList.AddRange(shipItemList);
+            itemsList.AddRange(rootItemList);
+            itemsList.AddRange(propItemsList);
+            //Remove the items from the banned list so so it should not be saved
+            itemsList.RemoveAll(grabObj => _bannedItemsFromSaving.Contains(grabObj.itemProperties.itemName) || _bannedItemsFromSaving.Contains(grabObj.itemProperties.name));
+            return itemsList.ToArray();
         }
         #endregion